+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.4.22
 Bug fix:
+- [#1671](https://github.com/elysiajs/elysia/issues/1671) mount() produces incorrect URL path when Elysia instance has prefix option
 - ValueError with summary missing types
 - Elysia not using Bun.routes